+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/order/infra/RedisOrderMessagePublisher.java
@@ -1,5 +1,7 @@
 package com.connectruck.foodtruck.order.infra;
 
+import com.connectruck.foodtruck.order.message.OrderCreatedMessage;
+import com.connectruck.foodtruck.order.message.OrderMessagePublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
@@ -7,7 +9,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class RedisOrderMessagePublisher implements OrderMessagePublisher {
+class RedisOrderMessagePublisher implements OrderMessagePublisher {
 
     private final RedisTemplate<String, OrderCreatedMessage> orderCreatedMessageTemplate;
     private final ChannelTopic orderCreatedChannelTopic;
