+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/order/message/OrderCreatedMessage.java
@@ -1,4 +1,4 @@
-package com.connectruck.foodtruck.order.infra;
+package com.connectruck.foodtruck.order.message;
 
 public record OrderCreatedMessage(long truckId, long orderId) {
 }
