+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/order/service/OrderService.java
@@ -23,8 +23,8 @@ import com.connectruck.foodtruck.order.dto.OrdererInfoRequest;
 import com.connectruck.foodtruck.order.dto.OrdersResponse;
 import com.connectruck.foodtruck.order.exception.OrderCreationException;
 import com.connectruck.foodtruck.order.exception.WrongOrderInfoException;
-import com.connectruck.foodtruck.order.infra.OrderCreatedMessage;
-import com.connectruck.foodtruck.order.infra.OrderMessagePublisher;
+import com.connectruck.foodtruck.order.message.OrderCreatedMessage;
+import com.connectruck.foodtruck.order.message.OrderMessagePublisher;
 import com.connectruck.foodtruck.truck.domain.Truck;
 import com.connectruck.foodtruck.truck.domain.TruckRepository;
 import java.time.LocalDateTime;
